user_service/requirements.txt
django==5.1.4
djangorestframework==3.15.2
redis==5.2.1
gunicorn==23.0.0